requirements.txt
absl-py==0.9.0
bleach==1.5.0
click==7.1.2
Flask==1.1.2
h5py==2.10.0
html5lib==0.9999999
importlib-metadata==1.6.0
itsdangerous==1.1.0
Jinja2==2.11.2
Keras==2.1.5
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
Markdown==3.2.2
MarkupSafe==1.1.1
numpy==1.18.4
opencv-python==4.2.0.34
pickle-mixin==1.0.2
Pillow==7.1.2
protobuf==3.12.1
PyYAML==5.3.1
scipy==1.4.1
six==1.15.0
tensorflow==1.5.0 
tensorflow-tensorboard==1.5.1
Werkzeug==1.0.1
zipp==3.1.0


gunicorn==20.0.0

